Restrict when startAtOperationTime is included for change streams
When resuming a change stream, the driver is required to save the
operationTime returned in the aggregate command response and send
it on subsequent aggregate commands when resuming the change stream,
until a resumeToken is available.

This commit ensures that it's only done when connected to a MongoDB 4.0
server, as per specification.

JAVA-2892