Initial API design for Connection to handle document sequences in the reply.
This could be a lot prettier.  In particular, it would be a more
sensible design if document sequences were abstracted below Connection
rather than bleeding into the Connection API.  One way to do this would
be to wrap a BsonBinaryReader in a BsonReader implementation that
understands document sequences and can insert them into the main
document at the appropriate point.  Something like:

DocumentSequenceAppenedingBsonReader implements BsonReader {
   BsonReader mainDocumentReader;
   ByteBuf documentSequence;
}

and pass that BsonReader to the Decoder, which can work the same as it
always has and is unaware of the existence of the document sequence.

It's possible though that this will add a performance penalty that the
feature was intended to address, and a less elegant design (as it is in
this patch) will be better.

PM-948