Implement connection pooling and monitoring specification
https://github.com/mongodb/specifications/tree/master/source/connection-monitoring-and-pooling

This implementation excludes events in the specification which will require
adding new methods to the ConnectionPoolListener interface.  New events
are deferred to the 4.0 release, when we can take advantage of Java 8
default methods to add new methods to an interface without breaking binary
compatibility. This work will be done in scope of JAVA-3150.

JAVA-3068